Screen Changes
I am trying to learn machScreen so there are lots of errors in these
screens.  I do not know how to create a user defined function so there
are many errors for that.

I don't know if the user will need to change the infill often so these
controls may want to be deleted.  I don't even have my SX3 CNC mill
converted to do 3D Printing yet so don't have a feel for this.

You can see I made it such that the user can *choose* to update their
config.ini file to the temp values they set or not.  I agree having them
have to chage the values each time is bad but forceing a config.ini
change is not good either.  I also made one of the buttons change the
fields back to what is in the config.ini file.

I now wonder if we have too many screens so I will commit this version
and then consolidate screens you can use whichever you want