Suggested Screens
Consider this screen set which would open to the main screen where all
of the config and main functions reside.  An About button takes the user
to the "Info" screen.

The screens depict a rough layout only where buttons, labels, etc, need
to be properly sized and position as well as set up as user defined
functions. I do not know how to do frames or the multiple screens nor do
I know how to to get these screens to show up in Mach3  and especially
what they will look like.  Also, I still am not sure the infill controls
are needed (i.e. how often do these change?)  Since there are no
checkboxes or radio buttons available the infill patter would show the
appropriate LED lit for the selected pattern.

I think the Gcode panel would populate when the Post to Mach3 button is
pressed there is nothing to really do in this regard and the edit button
would just shell out to notepad or whatever the user had set up in Mach3
for the Gcode editor.  Here again I don't know if seeing the Gcode
and/or being able to edit it here in the wizard is needed.  All I know
is that whenever i use Cut2D for a machined part i always end up
tweaking the Gcode.

Please see what you think of the of this approach and if you decide to
use it make any changes you wish to or are required to make to get it
functional.  We can then start to do the script writing.   OK?